+++ shared/src/commonMain/kotlin/shiny/mc/core/coordinators/transaction/AddRecordTransaction.kt
@@ -3,7 +3,7 @@ package shiny.mc.core.coordinators.transaction
 interface AddRecordTransaction {
     suspend fun addTransaction(
         recordId: String,
-        value: Float,
+        value: Double,
         purpose: String,
         datetime: Long,
     )
